+++ NEWS.md
@@ -1,3 +1,10 @@
+# common 0.0.9001
+
+* Add news.
+* Add readment.
+* Complete documentation.
+* Build pkgdown site.
+
 # common 0.0.9000
 
 * Initial version with following functions: 
@@ -6,4 +13,5 @@
 - `%p%`: An infix operator for the `paste0()` function.
 - `%eq`: An infix operator to determine equality between two objects.
 - `roundup()`: A function that rounds 5 away from zero.
+- `Sys.path()`: A function to get the path of the currently executing program.
 - `v()`: A quoting function for non-standard evaluation.